refactor: filter.ts 책임 분리 리펙토링
