+++ src/entities/domain/types.ts
@@ -1,0 +1,3 @@
+import type { Option } from '@shared/types/common';
+
+export type DomainResponse = Option;
