+++ src/pages/members/mocks/mock-skill-options.ts
@@ -1,5 +1,5 @@
 import type { SuccessResponse } from '@shared/api/types';
-import type { Option } from '@shared/types/filter/filter';
+import type { Option } from '@shared/types/common';
 
 export interface FieldSkillItemResponse {
   field: Option;
