+++ src/pages/post-create/mocks/mock-post-create-options.ts
@@ -1,9 +1,6 @@
-export interface PostFormOption {
-  id: number;
-  name: string;
-}
+import type { Option } from '@shared/types/common';
 
-export const MOCK_DOMAIN_OPTIONS: PostFormOption[] = [
+export const MOCK_DOMAIN_OPTIONS: Option[] = [
   { id: 1, name: 'IT/개발' },
   { id: 2, name: '디자인' },
   { id: 3, name: '기획' },
