+++ src/shared/types/common.ts
@@ -1,0 +1,4 @@
+export interface Option {
+  id: number;
+  name: string;
+}
